Improve handling of Meta Data from Assets (#3943)
