+++ bundles/AdminBundle/Controller/Admin/AssetController.php
@@ -120,41 +120,11 @@ class AssetController extends ElementControllerBase implements EventedController
                 $imageInfo['dimensions']['height'] = $asset->getHeight();
             }
 
-            $exifData = $asset->getEXIFData();
-            if (!empty($exifData)) {
-                $imageInfo['exif'] = $exifData;
-            }
-
-            $xmpData = $asset->getXMPData();
-            if (!empty($xmpData)) {
-                // flatten to a one-dimensional array
-                array_walk($xmpData, function (&$value) {
-                    if (is_array($value)) {
-                        $value = implode_recursive($value, ' | ');
-                    }
-                });
-                $imageInfo['xmp'] = $xmpData;
-            }
-
-            // check for VR meta-data
-            $mergedMetaData = array_merge($exifData, $xmpData);
-            if (isset($mergedMetaData['ProjectionType']) && $mergedMetaData['ProjectionType'] == 'equirectangular') {
-                $imageInfo['isVrImage'] = true;
-            }
-
-            $iptcData = $asset->getIPTCData();
-            if (!empty($iptcData)) {
-                // flatten data, to be displayed in grid
-                foreach ($iptcData as &$value) {
-                    if (is_array($value)) {
-                        $value = implode(', ', $value);
-                    }
-                }
-
-                $imageInfo['iptc'] = $iptcData;
-            }
-
             $imageInfo['exiftoolAvailable'] = (bool) \Pimcore\Tool\Console::getExecutable('exiftool');
+
+            if(!$asset->getEmbeddedMetaData(false)){
+                $asset->getEmbeddedMetaData(true, false); // read Exif, IPTC and XPM like in the old days ...
+            }
 
             $asset->imageInfo = $imageInfo;
         }
