+++ bundles/AdminBundle/Resources/public/js/pimcore/asset/image.js
@@ -37,6 +37,7 @@ pimcore.asset.image = Class.create(pimcore.asset.asset, {
         this.tagAssignment = new pimcore.element.tag.assignment(this, "asset");
         this.metadata = new pimcore.asset.metadata(this);
         this.workflows = new pimcore.element.workflows(this, "asset");
+        this.embeddedMetaData = new pimcore.asset.embedded_meta_data(this);
 
         this.getData();
     },
@@ -52,9 +53,9 @@ pimcore.asset.image = Class.create(pimcore.asset.asset, {
             items.push(this.getEditPanel());
         }
 
-        var exifPanel = this.getExifPanel();
-        if(exifPanel) {
-            items.push(exifPanel);
+        var embeddedMetaDataPanel = this.embeddedMetaData.getPanel();
+        if(embeddedMetaDataPanel) {
+            items.push(embeddedMetaDataPanel);
         }
 
         if (this.isAllowed("publish")) {
@@ -120,54 +121,6 @@ pimcore.asset.image = Class.create(pimcore.asset.asset, {
         return this.editPanel;
     },
 
-    getExifPanel: function () {
-        if (!this.exifPanel) {
-
-            if(!this.data["imageInfo"] || (!this.data["imageInfo"]["exif"] && !this.data["imageInfo"]["iptc"] && !this.data["imageInfo"]["xmp"])) {
-                return false;
-            }
-
-            var exifPanel = new Ext.grid.PropertyGrid({
-                title: 'EXIF',
-                flex: 1,
-                border: true,
-                source: this.data["imageInfo"]["exif"] || [],
-                clicksToEdit: 1000
-            });
-            exifPanel.plugins[0].disable();
-
-            var iptcPanel = new Ext.grid.PropertyGrid({
-                title: 'IPTC',
-                flex: 1,
-                border: true,
-                source: this.data["imageInfo"]["iptc"] || [],
-                clicksToEdit: 1000
-            });
-            iptcPanel.plugins[0].disable();
-
-            var xmpPanel = new Ext.grid.PropertyGrid({
-                title: 'XMP',
-                flex: 1,
-                border: true,
-                source: this.data["imageInfo"]["xmp"] || [],
-                clicksToEdit: 1000
-            });
-            xmpPanel.plugins[0].disable();
-
-            this.exifPanel = new Ext.Panel({
-                title: "EXIF/XMP/IPTC",
-                layout: {
-                    type: 'hbox',
-                    align: 'stretch'
-                },
-                iconCls: "pimcore_icon_exif",
-                items: [exifPanel, xmpPanel, iptcPanel]
-            });
-        }
-
-        return this.exifPanel;
-    },
-
     getDisplayPanel: function () {
 
         if (!this.displayPanel) {
@@ -229,9 +182,13 @@ pimcore.asset.image = Class.create(pimcore.asset.asset, {
 
             if (this.data.imageInfo.dimensions) {
 
+                var dimensions = [];
+                dimensions[t("width")] = this.data.imageInfo.dimensions.width;
+                dimensions[t("height")] = this.data.imageInfo.dimensions.height;
+
                 var dimensionPanel = new Ext.create('Ext.grid.property.Grid', {
                     title: t("details"),
-                    source: this.data.imageInfo.dimensions,
+                    source: dimensions,
                     autoHeight: true,
 
                     clicksToEdit: 1000,
