+++ bundles/AdminBundle/Resources/public/js/pimcore/asset/video.js
@@ -37,6 +37,7 @@ pimcore.asset.video = Class.create(pimcore.asset.asset, {
         this.tagAssignment = new pimcore.element.tag.assignment(this, "asset");
         this.metadata = new pimcore.asset.metadata(this);
         this.workflows = new pimcore.element.workflows(this, "asset");
+        this.embeddedMetaData = new pimcore.asset.embedded_meta_data(this);
 
         this.getData();
     },
@@ -46,6 +47,11 @@ pimcore.asset.video = Class.create(pimcore.asset.asset, {
         var user = pimcore.globalmanager.get("user");
 
         items.push(this.getEditPanel());
+
+        var embeddedMetaDataPanel = this.embeddedMetaData.getPanel();
+        if(embeddedMetaDataPanel) {
+            items.push(embeddedMetaDataPanel);
+        }
 
         if (this.isAllowed("publish")) {
             items.push(this.metadata.getLayout());
@@ -111,118 +117,146 @@ pimcore.asset.video = Class.create(pimcore.asset.asset, {
 
             var date = new Date();
 
+            var detailsData = [];
+            if(this.data.customSettings.videoWidth){
+                detailsData[t("width")] = this.data.customSettings.videoWidth;
+            }
+            if(this.data.customSettings.videoHeight){
+                detailsData[t("height")] = this.data.customSettings.videoHeight;
+            }
+            if(this.data.customSettings.duration){
+                detailsData[t("duration")] = this.data.customSettings.duration;
+            }
+
+            var dimensionPanel = new Ext.create('Ext.grid.property.Grid', {
+                title: t("details"),
+                source: detailsData,//this.data.imageInfo.dimensions, // Todo
+                autoHeight: true,
+
+                clicksToEdit: 1000,
+                viewConfig: {
+                    forceFit: true,
+                    scrollOffset: 2
+                }
+            });
+            dimensionPanel.plugins[0].disable();
+            dimensionPanel.getStore().sort("name", "DESC");
+
             this.previewImagePanel = new Ext.Panel({
                 width: 300,
                 region: "east",
                 scrollable: 'y',
-                items: [{
-                    title: t("tools"),
-                    bodyStyle: "padding: 10px;",
-                    items: [{
-                        xtype: "button",
-                        text: t("standard_preview"),
-                        iconCls: "pimcore_icon_image",
-                        width: "100%",
-                        textAlign: "left",
-                        style: "margin-top: 5px",
-                        handler: function () {
-                            if (this.previewMode != 'video') {
-                                this.initPreviewVideo();
-                            }
-                        }.bind(this)
-                    }, {
-                        xtype: "button",
-                        text: t("360_viewer"),
-                        iconCls: "pimcore_icon_vr",
-                        width: "100%",
-                        textAlign: "left",
-                        style: "margin-top: 5px",
-                        hidden: !(this.data['videoInfo'] && this.data['videoInfo']['previewUrl']),
-                        handler: function () {
-                            this.initPreviewVr();
-                        }.bind(this)
-                    }]
-                }, {
-                    xtype: "panel",
-                    title: t("select_image_preview"),
-                    height: 400,
-                    bodyStyle: "padding:10px",
-                    itemId: "inner",
-                    hidden: true,
-                    items: [{
-                        xtype: "container",
-                        style: "margin: 10px 0 10px 0;",
-                        height: 200,
-                        id: "pimcore_asset_video_imagepreview_" + this.id,
-                        html: '<img class="pimcore_video_preview_image" align="center" src="/admin/asset/get-video-thumbnail?id='
-                            + this.id + '&width=265&aspectratio=true&_dc=' + date.getTime() + '" />'
-                    }, {
-                        xtype: "button",
-                        text: t("use_current_player_position_as_preview"),
-                        iconCls: "pimcore_icon_video pimcore_icon_overlay_edit",
-                        width: "100%",
-                        handler: function () {
-                            try {
-                                this.previewImagePanel.getComponent("inner").getComponent("assetPath").setValue("");
-
-                                var time = window[this.previewFrameId].document.getElementById("video").currentTime;
-                                var date = new Date();
-                                var cmp = Ext.getCmp("pimcore_asset_video_imagepreview_" + this.id);
-                                cmp.update('<img class="pimcore_video_preview_image" align="center" src="/admin/asset/get-video-thumbnail?id='
-                                    + this.id + '&width=265&aspectratio=true&time=' + time + '&settime=true&_dc='
-                                    + date.getTime() + '" />');
-
-                            } catch (e) {
-                                console.log(e);
-                            }
-                        }.bind(this)
-                    }, {
-                        xtype: "container",
-                        border: false,
-                        style: "padding: 10px 0 10px 0;",
-                        html: t("or_specify_an_asset_image_below") + ":"
-                    }, {
-                        xtype: "textfield",
-                        itemId: "assetPath",
-                        fieldCls: "input_drop_target",
-                        width: "100%",
-                        listeners: {
-                            "render": function (el) {
-                                new Ext.dd.DropZone(el.getEl(), {
-                                    reference: el,
-                                    ddGroup: "element",
-                                    getTargetFromEvent: function (e) {
-                                        return this.getEl();
-                                    }.bind(el),
-
-                                    onNodeOver: function (target, dd, e, data) {
-                                        if (data.records.length == 1 && data.records[0].data.elementType == "asset") {
-                                            return Ext.dd.DropZone.prototype.dropAllowed;
-                                        }
-                                    },
-
-                                    onNodeDrop: function (el, target, dd, e, data) {
-                                        if (pimcore.helpers.dragAndDropValidateSingleItem(data)) {
-                                            data = data.records[0].data;
-                                            if (data.elementType == "asset") {
-                                                el.setValue(data.path);
-
-                                                var date = new Date();
-                                                var cmp = Ext.getCmp("pimcore_asset_video_imagepreview_" + this.id);
-                                                cmp.update('<img align="center" src="/admin/asset/get-video-thumbnail?id='
-                                                    + this.id + '&width=265&aspectratio=true&image='
-                                                    + data.id + '&setimage=true&_dc='
-                                                    + date.getTime() + '" />');
-                                                return true;
-                                            }
-                                        }
-                                        return false;
-                                    }.bind(this, el)
-                                });
+                items: [
+                    {
+                        title: t("tools"),
+                        bodyStyle: "padding: 10px;",
+                        items: [{
+                            xtype: "button",
+                            text: t("standard_preview"),
+                            iconCls: "pimcore_icon_image",
+                            width: "100%",
+                            textAlign: "left",
+                            style: "margin-top: 5px",
+                            handler: function () {
+                                if (this.previewMode != 'video') {
+                                    this.initPreviewVideo();
+                                }
                             }.bind(this)
-                        }
+                        }, {
+                            xtype: "button",
+                            text: t("360_viewer"),
+                            iconCls: "pimcore_icon_vr",
+                            width: "100%",
+                            textAlign: "left",
+                            style: "margin-top: 5px",
+                            hidden: !(this.data['videoInfo'] && this.data['videoInfo']['previewUrl']),
+                            handler: function () {
+                                this.initPreviewVr();
+                            }.bind(this)
+                        }]
+                    },
+                    dimensionPanel,
+                    {
+                        xtype: "panel",
+                        title: t("select_image_preview"),
+                        height: 400,
+                        bodyStyle: "padding:10px",
+                        itemId: "inner",
+                        hidden: true,
+                        items: [{
+                            xtype: "container",
+                            style: "margin: 10px 0 10px 0;",
+                            height: 200,
+                            id: "pimcore_asset_video_imagepreview_" + this.id,
+                            html: '<img class="pimcore_video_preview_image" align="center" src="/admin/asset/get-video-thumbnail?id='
+                                + this.id + '&width=265&aspectratio=true&_dc=' + date.getTime() + '" />'
+                        }, {
+                            xtype: "button",
+                            text: t("use_current_player_position_as_preview"),
+                            iconCls: "pimcore_icon_video pimcore_icon_overlay_edit",
+                            width: "100%",
+                            handler: function () {
+                                try {
+                                    this.previewImagePanel.getComponent("inner").getComponent("assetPath").setValue("");
+
+                                    var time = window[this.previewFrameId].document.getElementById("video").currentTime;
+                                    var date = new Date();
+                                    var cmp = Ext.getCmp("pimcore_asset_video_imagepreview_" + this.id);
+                                    cmp.update('<img class="pimcore_video_preview_image" align="center" src="/admin/asset/get-video-thumbnail?id='
+                                        + this.id + '&width=265&aspectratio=true&time=' + time + '&settime=true&_dc='
+                                        + date.getTime() + '" />');
+
+                                } catch (e) {
+                                    console.log(e);
+                                }
+                            }.bind(this)
+                        }, {
+                            xtype: "container",
+                            border: false,
+                            style: "padding: 10px 0 10px 0;",
+                            html: t("or_specify_an_asset_image_below") + ":"
+                        }, {
+                            xtype: "textfield",
+                            itemId: "assetPath",
+                            fieldCls: "input_drop_target",
+                            width: "100%",
+                            listeners: {
+                                "render": function (el) {
+                                    new Ext.dd.DropZone(el.getEl(), {
+                                        reference: el,
+                                        ddGroup: "element",
+                                        getTargetFromEvent: function (e) {
+                                            return this.getEl();
+                                        }.bind(el),
+
+                                        onNodeOver: function (target, dd, e, data) {
+                                            if (data.records.length == 1 && data.records[0].data.elementType == "asset") {
+                                                return Ext.dd.DropZone.prototype.dropAllowed;
+                                            }
+                                        },
+
+                                        onNodeDrop: function (el, target, dd, e, data) {
+                                            if (pimcore.helpers.dragAndDropValidateSingleItem(data)) {
+                                                data = data.records[0].data;
+                                                if (data.elementType == "asset") {
+                                                    el.setValue(data.path);
+
+                                                    var date = new Date();
+                                                    var cmp = Ext.getCmp("pimcore_asset_video_imagepreview_" + this.id);
+                                                    cmp.update('<img align="center" src="/admin/asset/get-video-thumbnail?id='
+                                                        + this.id + '&width=265&aspectratio=true&image='
+                                                        + data.id + '&setimage=true&_dc='
+                                                        + date.getTime() + '" />');
+                                                    return true;
+                                                }
+                                            }
+                                            return false;
+                                        }.bind(this, el)
+                                    });
+                                }.bind(this)
+                            }
+                        }]
                     }]
-                }]
             });
 
             this.previewImagePanel.on("beforedestroy", function () {
