+++ bundles/CoreBundle/Resources/translations/de.json
