+++ bundles/CoreBundle/Resources/translations/en.json
@@ -866,5 +866,7 @@
   "paste_recursive_as_language_variant": "Paste as new language variant (recursive)",
   "paste_no_new_language_error": "All system languages already linked to source element.",
   "reload_column_configuration": "Reload column configuration",
-  "reload_column_configuration_notice": "Note: Reloading column configuration will reset selected grid columns"
+  "reload_column_configuration_notice": "Note: Reloading column configuration will reset selected grid columns",
+  "embedded_meta_data": "Embedded Meta Info",
+  "duration": "Duration"
 }
